neatening of architectures
